fix: gocyclo cyclomatic complexity
Gocyclo calculates cyclomatic complexities of functions in Go source code. The cyclomatic complexity of a function is calculated according to the following rules: 1 is the base complexity of a function +1 for each 'if', 'for', 'case', '&&' or '||' Go Report Card warns on functions with cyclomatic complexity > 15.

framework/model/helper/helpers.go
Line 21: warning: cyclomatic complexity 16 of function fillStruct() is high (> 15) (gocyclo)

Signed-off-by: Toby Yan <me@tobyan.com>